+++ cluster/pulumi/canton-network/src/svc.ts
@@ -140,6 +140,12 @@ export class Svc extends pulumi.ComponentResource {
     }));
 
     const foundingSvRewardWeightBps = 140_000;
+    const initialTickDurationS = 600;
+    // chosen for $1/360 days at above tick duration
+    const initialHoldingFee =
+      Math.round(
+        (100000_00000 * 1.0) / 360.0 / ((24.0 * 60.0 * 60.0) / initialTickDurationS) / 0.005 // assumed coin price
+      ) / 100000_00000;
 
     const runningMigration = this.args.globalDomainUpgradeConfig.isRunningMigration();
     const founder = await this.installSvNode(
@@ -150,6 +156,8 @@ export class Svc extends pulumi.ComponentResource {
             type: 'found-collective',
             foundingSvRewardWeightBps,
             roundZeroDuration: process.env.ROUND_ZERO_DURATION,
+            initialTickDurationS,
+            initialHoldingFee,
           },
       {
         founder: founderCometBftConf,